+++ auth-db/queries/email-auth/index.ts
@@ -2,3 +2,4 @@ export * from "./create.ts";
 export * from "./get-by-email.ts";
 export * from "./update-verification.ts";
 export * from "./update-forgot-password-token.ts";
+export * from "./update-password-hash.ts";
